requirements.txt
# Physics-Guided Ignition Probability Modelling
# Core data stack
numpy>=1.24
pandas>=2.0
xarray>=2023.1
netCDF4>=1.6
scipy>=1.10

# Geospatial
rasterio>=1.3
geopandas>=0.13
shapely>=2.0
pyproj>=3.5

# Machine Learning
scikit-learn>=1.3

# Visualization
matplotlib>=3.7
seaborn>=0.12
cartopy>=0.21

# Data access APIs
cdsapi>=0.6
requests>=2.28

# Jupyter
jupyter>=1.0
notebook>=7.0
ipykernel>=6.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65

# Advanced ML
xgboost>=2.0
shap>=0.42

# Dashboard
streamlit>=1.30
plotly>=5.18
folium>=0.15
streamlit-folium>=0.15

# Optional: for larger-than-memory datasets
# dask>=2023.1
